+++ changelog.md
@@ -25,4 +25,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.1.0] 2020-12-30
 
-- Add cmd/eng for querying the english language
+- Add cmd/eng for querying the English language